requirements.txt
django-tastypie>=0.9.14
django-registration>=0.8
